kernel refactor
