+++ kernel/src/arch/x86/mmu/VMM.c
@@ -1,5 +1,5 @@
-#include <cpu/mmu/VMM.h>
-#include <cpu/mmu/PMM.h>
+#include <arch/x86/mmu/VMM.h>
+#include <arch/x86/mmu/PMM.h>
 #include <lib/string.h>
 #include <defs/interrupts.h>
 #include <lib/ISR_IRQ.h>
